+++ Modules/Shop/Routes/web.php
@@ -15,6 +15,7 @@ Route::group(['middleware' => ['auth', 'language']], function () {
     Route::prefix('shop')->name('shop.')->middleware(['can:validate-shop-coupons'])->group(function(){
         Route::get('/', 'ShopController@index')->name('index');
         Route::post('/', 'ShopController@redeem')->name('redeem');
+        Route::post('/cancelCard', 'ShopController@cancelCard')->name('cancelCard');
         Route::get('/settings', 'ShopSettingsController@edit')->name('settings.edit')->middleware(['can:configure-shop']);
         Route::put('/settings', 'ShopSettingsController@update')->name('settings.update')->middleware(['can:configure-shop']);
     });